+++ experimental/packages/exporter-trace-otlp-grpc/test/OTLPTraceExporter.test.ts
@@ -5,8 +5,9 @@
 
 import { OTLPTraceExporter } from '../src';
 import type { ServerTestContext } from './utils';
-import { startServer } from './utils';
+import { startServer, TestMetricReader } from './utils';
 import * as assert from 'assert';
+import { MeterProvider } from '@opentelemetry/sdk-metrics';
 import { SimpleSpanProcessor, TracerProvider } from '@opentelemetry/sdk-trace';
 
 const testServiceDefinition = {
@@ -65,11 +66,16 @@ describe('OTLPTraceExporter', function () {
 
   it('successfully exports data', async () => {
     // arrange
+    const metricReader = new TestMetricReader();
+    const meterProvider = new MeterProvider({
+      readers: [metricReader],
+    });
     const tracerProvider = new TracerProvider({
       spanProcessors: [
         new SimpleSpanProcessor({
           exporter: new OTLPTraceExporter({
             url: 'http://localhost:1501',
+            selfObsMeterProvider: meterProvider,
           }),
         }),
       ],
@@ -81,5 +87,12 @@ describe('OTLPTraceExporter', function () {
 
     // assert
     assert.strictEqual(serverTestContext.requests.length, 1);
+
+    const metrics = await metricReader.collect();
+    const scopeMetrics = metrics.resourceMetrics.scopeMetrics.find(
+      sm => sm.scope.name === '@opentelemetry/otlp-exporter'
+    );
+    assert.ok(scopeMetrics);
+    await meterProvider.shutdown();
   });
 });