+++ experimental/packages/otlp-grpc-exporter-base/src/index.ts
@@ -4,7 +4,10 @@
  */
 
 export { convertLegacyOtlpGrpcOptions } from './configuration/convert-legacy-otlp-grpc-options';
-export { createOtlpGrpcExportDelegate } from './otlp-grpc-export-delegate';
+export {
+  createOtlpGrpcExportDelegate,
+  createOtlpGrpcExporterMetrics,
+} from './otlp-grpc-export-delegate';
 export {
   createEmptyMetadata,
   createInsecureCredentials,